fixed data types

+++ FebLaunch/main.cpp
@@ -309,7 +309,7 @@ int main() {
 
         /////////////       READ CALIBRATION OF MS5607_1        /////////////
         //to test for data collisions with BMP388//
-        short C1, C2, C3, C4, C5, C6;
+        unsigned int C1, C2, C3, C4, C5, C6;
     ms5607_1.readPublicConfig(&C1,&C2,&C3,&C4,&C5,&C6);
     cout << C1 << " " << C2 << " " << C3 << " " << C4 << " " << C5 << " " << C6 << "\n\n";
